beryllium::events

Enum Event

pub enum Event {

    Quit,
    DisplayConnected {
        display_index: u32,
    },
    DisplayDisconnected {
        display_index: u32,
    },
    DisplayOrientationChanged {
        display_index: u32,
        new_orientation: DisplayOrientation,
    },
    WindowShown {
        win_id: u32,
    },
    WindowHidden {
        win_id: u32,
    },
    WindowExposed {
        win_id: u32,
    },
    WindowMoved {
        win_id: u32,
        x: i32,
        y: i32,
    },
    WindowResized {
        win_id: u32,
        width: i32,
        height: i32,
    },
    WindowSizeChanged {
        win_id: u32,
        width: i32,
        height: i32,
    },
    WindowMinimized {
        win_id: u32,
    },
    WindowMaximized {
        win_id: u32,
    },
    WindowRestored {
        win_id: u32,
    },
    MouseEnteredWindow {
        win_id: u32,
    },
    MouseExitedWindow {
        win_id: u32,
    },
    WindowGainedKeyboardFocus {
        win_id: u32,
    },
    WindowLostKeyboardFocus {
        win_id: u32,
    },
    WindowCloseRequest {
        win_id: u32,
    },
    Key {
        win_id: u32,
        pressed: bool,
        repeat: u8,
        scancode: SDL_Scancode,
        keycode: SDL_Keycode,
        modifiers: SDL_Keymod,
    },
    TextInput {
        win_id: u32,
        text: String,
    },
    MouseMotion {
        win_id: u32,
        mouse_id: u32,
        button_state: u32,
        x_win: i32,
        y_win: i32,
        x_delta: i32,
        y_delta: i32,
    },
    MouseButton {
        win_id: u32,
        mouse_id: u32,
        button: u8,
        pressed: bool,
        clicks: u8,
        x: i32,
        y: i32,
    },
    MouseWheel {
        win_id: u32,
        mouse_id: u32,
        x: i32,
        y: i32,
    },
    JoystickAxis {
        joy_id: i32,
        axis: u8,
        value: i16,
    },
    JoystickBall {
        joy_id: i32,
        ball: u8,
        x_rel: i16,
        y_rel: i16,
    },
    JoystickHat {
        joy_id: i32,
        hat: u8,
        value: u8,
    },
    JoystickButton {
        joy_id: i32,
        button: u8,
        pressed: bool,
    },
    JoystickAdded {
        index: i32,
    },
    JoystickRemoved {
        joy_id: i32,
    },
    ControllerAxis {
        ctrl_id: i32,
        axis: ControllerAxis,
        value: i16,
    },
    ControllerButton {
        ctrl_id: i32,
        button: ControllerButton,
        pressed: bool,
    },
    ControllerAdded {
        index: i32,
    },
    ControllerRemoved {
        ctrl_id: i32,
    },
    ControllerRemapped {
        ctrl_id: i32,
    },
    ControllerSensor {
        ctrl_id: i32,
        sensor: i32,
        data: [f32; 3],
    },
    AudioDeviceAdded {
        index: u32,
        is_capture: bool,
    },
    AudioDeviceRemoved {
        audio_id: u32,
        is_capture: bool,
    },
    Sensor {
        sensor_id: i32,
        data: [f32; 6],
    },
    DropBegin {
        win_id: u32,
    },
    DropFile {
        win_id: u32,
        name: String,
    },
    DropComplete {
        win_id: u32,
    },

}

Variants

Quit

DisplayConnected

Fields

display_index: u32

DisplayDisconnected

Fields

display_index: u32

DisplayOrientationChanged

Fields

display_index: u32

new_orientation: DisplayOrientation

WindowShown

Fields

win_id: u32

WindowHidden

Fields

win_id: u32

WindowExposed

Fields

win_id: u32

WindowMoved

Fields

win_id: u32

x: i32

y: i32

WindowResized

Fields

win_id: u32

width: i32

height: i32

WindowSizeChanged

Fields

win_id: u32

width: i32

height: i32

WindowMinimized

Fields

win_id: u32

WindowMaximized

Fields

win_id: u32

WindowRestored

Fields

win_id: u32

MouseEnteredWindow

Fields

win_id: u32

MouseExitedWindow

Fields

win_id: u32

WindowGainedKeyboardFocus

Fields

win_id: u32

WindowLostKeyboardFocus

Fields

win_id: u32

WindowCloseRequest

Fields

win_id: u32

Key

Fields

win_id: u32

pressed: bool

repeat: u8

scancode: SDL_Scancode

keycode: SDL_Keycode

modifiers: SDL_Keymod

TextInput

Fields

win_id: u32

text: String

MouseMotion

Mouse cursor motion

x_win and y_win are the window-relative mouse position.

x_delta and y_delta are the change in position since the last event.

button_state has bit N set when mouse button N is held down during the event.

Fields

win_id: u32

mouse_id: u32

button_state: u32

x_win: i32

y_win: i32

x_delta: i32

y_delta: i32

MouseButton

Fields

win_id: u32

mouse_id: u32

button: u8

pressed: bool

clicks: u8

x: i32

y: i32

MouseWheel

Mouse wheel change

x: horizontal, with positive to the right.

y: vertical, with positive away from the user.

Fields

win_id: u32

mouse_id: u32

x: i32

y: i32

JoystickAxis

Fields

joy_id: i32

axis: u8

value: i16

JoystickBall

Fields

joy_id: i32

ball: u8

x_rel: i16

y_rel: i16

JoystickHat

Fields

joy_id: i32

hat: u8

value: u8

JoystickButton

Fields

joy_id: i32

button: u8

pressed: bool

JoystickAdded

Fields

index: i32

JoystickRemoved

Fields

joy_id: i32

ControllerAxis

Fields

ctrl_id: i32

axis: ControllerAxis

value: i16

ControllerButton

Fields

ctrl_id: i32

button: ControllerButton

pressed: bool

ControllerAdded

Fields

index: i32

ControllerRemoved

Fields

ctrl_id: i32

ControllerRemapped

Fields

ctrl_id: i32

ControllerSensor

Fields

ctrl_id: i32

sensor: i32

data: [f32; 3]

AudioDeviceAdded

Fields

index: u32

is_capture: bool

AudioDeviceRemoved

Fields

audio_id: u32

is_capture: bool

Sensor

Fields

sensor_id: i32

data: [f32; 6]

DropBegin

Marks the start of a series of files being dropped onto the window.

Fields

win_id: u32

DropFile

The name of a file or directory the user dropped into the window.

Fields

win_id: u32

name: String

DropComplete

This marks the end of a group of file drops.

Fields

win_id: u32